mspack - removing ssize_t typing to ensure VS15 compilation

+++ libclamav/libmspack-0.5alpha/mspack/cabd.c
@@ -522,7 +522,7 @@ static char *cabd_read_string(struct mspack_system *sys,
   off_t base = sys->tell(fh);
   char buf[256], *str;
   unsigned int i, ok;
-  ssize_t len;
+  int len;
 
   /* read up to 256 bytes */
   len = sys->read(fh, &buf[0], 256);
